segus_engineering_Backend/requirements-dev.txt
# Requirements pour le développement - étend requirements.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2
jupyter==1.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance
django-silk==5.0.4
memory-profiler==0.61.0

# Security
bandit==1.7.5
safety==2.3.5

# API Documentation
drf-spectacular==0.26.5

# Environment Management
python-decouple==3.8
